requirements_otel.txt
opentelemetry-instrumentation-aiohttp-client==0.39b0
opentelemetry-util-http==0.39b0
asgiref~=3.6
opentelemetry-instrumentation-asgi==0.39b0
opentelemetry-instrumentation-asyncpg==0.39b0
opentelemetry-instrumentation-boto==0.39b0
opentelemetry-instrumentation-botocore==0.39b0
opentelemetry-instrumentation-celery==0.39b0
opentelemetry-instrumentation-dbapi==0.39b0
opentelemetry-instrumentation-django==0.39b0
opentelemetry-instrumentation-elasticsearch==0.39b0
opentelemetry-instrumentation-fastapi==0.39b0
opentelemetry-instrumentation-falcon==0.39b0
opentelemetry-instrumentation-flask==0.39b0
opentelemetry-instrumentation-grpc==0.39b0
opentelemetry-instrumentation-jinja2==0.39b0
opentelemetry-instrumentation-mysql==0.39b0
opentelemetry-instrumentation-psycopg2==0.39b0
opentelemetry-instrumentation-pymemcache==0.39b0
opentelemetry-instrumentation-pymongo==0.39b0
opentelemetry-instrumentation-pymysql==0.39b0
opentelemetry-instrumentation-pyramid==0.39b0
opentelemetry-instrumentation-redis==0.39b0
opentelemetry-instrumentation-requests==0.39b0
opentelemetry-instrumentation-sqlalchemy==0.39b0
opentelemetry-instrumentation-sqlite3==0.39b0
opentelemetry-instrumentation-starlette==0.39b0
opentelemetry-instrumentation-tornado==0.39b0
opentelemetry-instrumentation-wsgi==0.39b0